+++ front/esbuild.worker.ts
@@ -1,4 +1,5 @@
 import esbuild from "esbuild";
+import fs from 'fs';
 
 async function buildWorker() {
   try {
@@ -22,6 +23,8 @@ async function buildWorker() {
     console.log(
       `📦 Bundle size: ${(result.metafile.outputs["dist/start_worker.js"].bytes / 1024 / 1024).toFixed(2)} MB`
     );
+
+    fs.writeFileSync('dist/meta.json', JSON.stringify(result.metafile))
 
     // Log any warnings
     if (result.warnings.length > 0) {
