+++ front/lib/api/content_nodes.ts
@@ -2,7 +2,7 @@
 // eslint-disable-next-line dust/enforce-client-types-in-public-api
 import { INTERNAL_MIME_TYPES } from "@dust-tt/client";
 
-import { SPREADSHEET_INTERNAL_MIME_TYPES } from "@app/lib/content_nodes";
+import { SPREADSHEET_INTERNAL_MIME_TYPES } from "@app/lib/content_nodes_constants";
 import type { DataSourceViewResource } from "@app/lib/resources/data_source_view_resource";
 import type {
   ContentNodesViewType,
